# Perf/fetch blockheader from peer on FCU if block so that it start beacon header sync.