--- a/src/Nethermind/Nethermind.Merge.Plugin.Test/EngineModuleTests.Synchronization.cs
+++ b/src/Nethermind/Nethermind.Merge.Plugin.Test/EngineModuleTests.Synchronization.cs
@@ -118,7 +118,7 @@ public partial class EngineModuleTests
             .GetBlockHeaders(Arg.Any<Hash256>(), Arg.Any<int>(), Arg.Any<int>(), Arg.Any<CancellationToken>())
             .Returns(Task.FromResult(new BlockHeader[1] { block.Header }));
         SyncPeerAllocation alloc = new SyncPeerAllocation(new PeerInfo(syncPeer), AllocationContexts.All);
-        alloc.AllocateBestPeer(new []{new PeerInfo(syncPeer)}, Substitute.For<INodeStatsManager>(), Substitute.For<IBlockTree>());
+        alloc.AllocateBestPeer(new[] { new PeerInfo(syncPeer) }, Substitute.For<INodeStatsManager>(), Substitute.For<IBlockTree>());
         chain.SyncPeerPool
             .Allocate(
                 Arg.Any<IPeerAllocationStrategy>(),