--- a/src/Nethermind/Nethermind.Synchronization/Peers/ISyncPeerPool.cs
+++ b/src/Nethermind/Nethermind.Synchronization/Peers/ISyncPeerPool.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Nethermind.Blockchain.Synchronization;
+using Nethermind.Core;
 using Nethermind.Core.Crypto;
 using Nethermind.Stats.Model;
 using Nethermind.Synchronization.Peers.AllocationStrategies;
@@ -18,13 +19,7 @@ namespace Nethermind.Synchronization.Peers
             IPeerAllocationStrategy peerAllocationStrategy,
             AllocationContexts allocationContexts,
             int timeoutMilliseconds = 0,
-            CancellationToken? cancellationToken = null);
-
-        Task<T> AllocateAndRun<T>(
-            Func<ISyncPeer, Task<T>> func,
-            IPeerAllocationStrategy peerAllocationStrategy,
-            AllocationContexts allocationContexts,
-            CancellationToken cancellationToken);
+            CancellationToken cancellationToken = default);
 
         void Free(SyncPeerAllocation syncPeerAllocation);
 
@@ -106,5 +101,44 @@ namespace Nethermind.Synchronization.Peers
         event EventHandler<PeerBlockNotificationEventArgs> NotifyPeerBlock;
 
         event EventHandler<PeerHeadRefreshedEventArgs> PeerRefreshed;
+    }
+
+    public static class SyncPeerPoolExtensions
+    {
+        public static async Task<T> AllocateAndRun<T>(this ISyncPeerPool syncPeerPool, Func<ISyncPeer, Task<T>> func, IPeerAllocationStrategy peerAllocationStrategy, AllocationContexts allocationContexts,
+            CancellationToken cancellationToken)
+        {
+            SyncPeerAllocation? allocation = await syncPeerPool.Allocate(peerAllocationStrategy, allocationContexts, cancellationToken: cancellationToken);
+            try
+            {
+                if (allocation is null) return default;
+                return await func(allocation.Current?.SyncPeer);
+            }
+            finally
+            {
+                syncPeerPool.Free(allocation);
+            }
+        }
+
+
+        public static async Task<BlockHeader?> FetchHeaderFromPeer(this ISyncPeerPool syncPeerPool, Hash256 hash, CancellationToken cancellationToken)
+        {
+            BlockHeader[]? headers = null;
+            try
+            {
+                headers = await syncPeerPool.AllocateAndRun(
+                    peer => peer.GetBlockHeaders(hash, 1, 0, cancellationToken),
+                    BySpeedStrategy.FastestHeader,
+                    AllocationContexts.Headers,
+                    cancellationToken);
+            }
+            catch (OperationCanceledException)
+            {
+                // Timeout or no peer.
+                return null;
+            }
+
+            return headers?.Length == 1 ? headers[0] : null;
+        }
     }
 }