@asdacap 2024-02-02
Its not you. In the first graph, lighthouse losses all peer in the middle of the run, so it hangs for a few minute. Total time goes from about 32minute down to 21minute. In the second graph, it take about 19 minute. Does not improve much. Could be that the blocking FCU contribute to slowdown in lighthouse.