@asdacap 2024-02-02
It never reach 9 block per second consistently for some reason. Lighthouse does not sync fast enough.