@asdacap 2024-02-02
I suppose in the first graph, you can see lighthouse take about 16 minute to catch up. In the second graph, it probably take about the same amount of time, and nethermind take additional 2 minute to fully catch up. Not sure why it needed that though.